requirements.txt
geopandas
requests
pydantic
pyhumps
matplotlib